pyproject.toml
[project]
name = "rerun-realsense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcap>=1.3.1",
    "numpy>=2.2.6",
    "pynput>=1.8.1",
    "pyrealsense2>=2.56.5.9235",
    "pytest>=9.0.2",
    "rerun-sdk==0.21.0",
    "scipy>=1.15.3",
    "spectacularai[full]>=1.33.0",
    "zarr>=2.18.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"] # Error, Flake8, Isort (sorts imports)

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true # Useful for pyrealsense2 which lacks type stubs

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = "test_*.py"
